requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-test-plus==2.2.1
djangorestframework==3.14.0
freezegun==1.2.2
lxml==4.9.0
pyhwp==0.1b15
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
